// src/app/types/enums/repoField.enum.ts
export enum RepoFieldsEnum {
  NAME = 'name',
  DESCRIPTION = 'description',
  LANGUAGE = 'language',
  STARGAZERS_COUNT = 'stargazers_count',
  FORKS_COUNT = 'forks_count',
  OPEN_ISSUES = 'open_issues',
  CREATED_AT = 'created_at',
  UPDATED_AT = 'updated_at',
  OWNER_LOGIN = 'owner_login',
  AVATAR_URL = 'avatar_url',
}